Test repeatability
