+++ skema/text_reading/mention_linking/gromet_linker/uid_stamper.py
@@ -1,0 +1,16 @@
+
+
+class UidStamper():
+	def stamp(self, doc_id: str) -> str:
+		pass
+
+class DocIdStamper(UidStamper):
+	def stamp(self, doc_id: str) -> str:
+		return doc_id
+
+
+class HashStamper(UidStamper):
+	def stamp(self, doc_id: str) -> str:
+		return str(hash(doc_id))
+
+uid_stamper = DocIdStamper()
